+++ Laser.Sources/Laser.Orchard/Modules/Laser.Orchard.Vimeo/Models/UploadsInProgressRecord.cs
@@ -6,8 +6,8 @@ using System.Web;
 
 namespace Laser.Orchard.Vimeo.Models {
     public class UploadsInProgressRecord {
-        public virtual int UploadSize { get; set; } //total size of the upload, in Bytes
-        public virtual int UploadedSize { get; set; } //number of Bytes we have uploaded already
+        public virtual Int64 UploadSize { get; set; } //total size of the upload, in Bytes
+        public virtual Int64 UploadedSize { get; set; } //number of Bytes we have uploaded already
         //Information from the response to the streaming upload request
         public virtual string Uri { get; set; } //API endpoint of the upload ticket
         public virtual string CompleteUri { get; set; } //URI to use when finishing the upload
